[Base-view-model] 1. Change Generic Class viewModel
