+++ app/src/main/java/com/nanamare/movie/ui/NavigationViewModel.kt
@@ -4,7 +4,7 @@ import com.nanamare.domain.model.GenreModel
 import com.nanamare.movie.model.Result
 import kotlinx.coroutines.flow.Flow
 
-interface NavigationViewModel {
+interface NavigationViewModel: BaseViewModel {
     fun navigate(screen: Screen)
     val navigation: Flow<Screen>
